mcp-scanner/pyproject.toml
[project]
name = "mcp-scanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "openai>=1.3.0",
    "requests>=2.31.0",
    "google-generativeai>=0.8.3",
    "outlines>=0.1.0",
    "pydantic>=2.0.0",
]
